refactor: extract session responses from AgentWorkspace
